backend/requirements.txt
pandas
scikit-learn
matplotlib
fastapi
uvicorn
python-dotenv
